Modified design here a bit; instead of sticking the option to collapse/not-collapse white space on the Detection element of the rule struct, I moved it to the Parser itself, which is who really cares about it at time of parsing and setup.
Basically, the path now is Config -> RuleHandle -> Parser for the
option; at the end of the day we need to know as we're setting up our
`Keyword` and `Selection` structs if we need to both collapse the rule
setting them up and collapse the data as it comes through (or not).

This is a linked setting - meaning if you tell the engine to collapse
whitespace, it collapses whitespace for both rules and data, if you tell
it not to, it does not.