Version 1 of placeholder functionality implementation as defined in: https://github.com/SigmaHQ/sigma/wiki/Specification#placeholders.
This particular implementation defines a new function as `type
PlaceholderLookup func(string) ([]string, bool)` which is passed to a
new interface function `MatchEx(Event, PlaceholderLookup)(bool, bool)`.
Which cascades through the various *Match* functions in the engine.

During parse time, we pullout placeholders (which can ONLY be
stand-alone, non-glob, non-regex strings) and mark them as such. This
enables the caller to then pass a placeholder lookup function (as
described above) to the new MatchEx call, allowing the engine to expand
placeholders in a reasonable way to match on.

For example, a placeholder may be defined as '%administrators%', which
is basically a list of all admin usernames. The engine would effectively
pass 'administrators' to the placeholder function and rip over the list
of returned strings, looking for a match.

If the PlaceholderLookup function is passed as `nil` OR the old
`Match` function is called instead of `MatchEx`, we will ignore the
placeholder lookup and treat it as a normal string compare (restoring
the '%' on either side during match time, mimicing current behavior).